# sdk/log: Remove slice allocation from SimpleProcessor.OnEmit